@@ -7,33 +7,34 @@
     "dev": "vite --open",
     "build": "vite build",
     "preview": "vite preview",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore",
+    "lint": "eslint . --fix",
     "format": "prettier --write src/"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.7.2",
-    "@modyfi/vite-plugin-yaml": "^1.1.0",
+    "@modyfi/vite-plugin-yaml": "^1.1.1",
     "@popperjs/core": "^2.11.8",
-    "bootstrap": "^5.3.3",
+    "bootstrap": "^5.3.5",
     "js-yaml-loader": "^1.2.2",
-    "marked": "^15.0.4",
+    "marked": "^15.0.8",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
   },
   "devDependencies": {
-    "@rushstack/eslint-patch": "^1.10.4",
-    "@vitejs/plugin-vue": "^5.2.1",
-    "@vue/eslint-config-prettier": "^10.1.0",
-    "eslint": "^9.17.0",
-    "eslint-plugin-vue": "^9.32.0",
-    "prettier": "^3.4.2",
-    "sass": "^1.83.0",
-    "vite": "^6.0.6",
-    "vite-plugin-vue-devtools": "^7.6.8"
+    "@eslint/js": "^9.24.0",
+    "@rushstack/eslint-patch": "^1.11.0",
+    "@vitejs/plugin-vue": "^5.2.3",
+    "@vue/eslint-config-prettier": "^10.2.0",
+    "eslint": "^9.24.0",
+    "eslint-plugin-vue": "^9.33.0",
+    "prettier": "^3.5.3",
+    "sass": "^1.86.3",
+    "vite": "^6.2.6",
+    "vite-plugin-vue-devtools": "^7.7.2"
   },
   "volta": {
-    "node": "22.9.0",
-    "yarn": "4.5.0"
+    "node": "22.14.0",
+    "yarn": "4.9.0"
   },
-  "packageManager": "yarn@4.5.0"
+  "packageManager": "yarn@4.9.0"
 }
