@@ -5,20 +5,18 @@
         <p class="mb-0 me-2">© {{ new Date().getFullYear() }}</p>
         <p class="mb-0 ps-2 border-start">
           Johan LOPES
-          <span class="mx-1 fw-light d-none d-md-inline">Concepteur de solutions web Freelance</span>
+          <span class="mx-1 fw-light d-none d-md-inline"
+            >Concepteur de solutions web Freelance</span
+          >
         </p>
 
-        <router-link :to="{ name: 'legal-notices' }" class="text-decoration-none ms-auto"> Mentions légales </router-link>
+        <router-link :to="{ name: 'legal-notices' }" class="text-decoration-none ms-auto">
+          Mentions légales
+        </router-link>
       </div>
     </div>
   </footer>
 </template>
-
-<script>
-export default {
-  name: 'BottomBar'
-}
-</script>
 
 <style lang="scss" scoped>
 #footer {
